basic principle of landing on doc guides in inside the app dist folder.  Now we properly listen via ipc when the app is started, pass the appPath to the store, and load the guides folder with an example hardocs project describing how to work with hardocs.
REQUIREMENT: This documents can be edited by the user but never saved, so the save option needs to be disabled for this particular case.
