+++ guides/.hardocs/hardocs.json
@@ -1,0 +1,9 @@
+{
+  "id": "2a644dd7-e456-4eba-8f34-6b2b1d87e068",
+  "path": "D:\\my-projects\\hardocs\\REPOS\\hardocs-vue-client",
+  "docsDir": "docs",
+  "entryFile": "index.md",
+  "name": "guides",
+  "shortTitle": "Hardocs getting started documentation",
+  "updatedAt": "new Date().toISOString()"
+}
