+++ guides/docs/index.md
@@ -1,0 +1,12 @@
+
+---
+title: Read the docs here
+description: Read the docs here
+---
+
+
+Read the docs here
+------------------
+
+Dont delete them
+
